+++ Turbo/src/Turbo/Renderer/VertexBuffer.h
@@ -13,7 +13,7 @@ namespace Turbo
         static Ref<VertexBuffer> Create(const VertexBuffer::Config& config);
         virtual ~VertexBuffer();
 
-        virtual void SetData(void* data, u32 size) = 0;
+        virtual void SetData(void* data, size_t size) = 0;
     protected:
         VertexBuffer(const VertexBuffer::Config& config);
